added exceptions to ReviewsService methods

+++ src/main/java/com/librarSystem/reviews/ReviewsService.java
@@ -1,5 +1,8 @@
 package com.librarSystem.reviews;
 
+import com.librarSystem.books.BooksService;
+import com.librarSystem.exception.ResourceNotFound;
+import com.librarSystem.users.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,29 +12,51 @@ import java.util.List;
 public class ReviewsService {
 
     private ReviewsDAO reviewsDAO;
+    private UsersService usersService;
+    private BooksService booksService;
 
     @Autowired
-    public ReviewsService(ReviewsDAO reviewsDAO) {
+    public ReviewsService(ReviewsDAO reviewsDAO, UsersService usersService, BooksService booksService) {
         this.reviewsDAO = reviewsDAO;
+        this.usersService = usersService;
+        this.booksService = booksService;
     }
 
     public List<Reviews> getReviewsByBook (int bookid){
+        if(reviewsDAO.getReviewsByBook(bookid).isEmpty()){
+            throw new ResourceNotFound("no reviews available");
+        }
         return reviewsDAO.getReviewsByBook(bookid);
     }
 
     public List<Reviews> getReviewsByUser (int userid){
+        if(reviewsDAO.getReviewsByUser(userid).isEmpty()){
+            throw new ResourceNotFound("no reviews found");
+        }
         return reviewsDAO.getReviewsByUser(userid);
     }
 
     public int addReview (int userid, int bookid, Reviews reviews){
+        if(usersService.selectUserById(userid).isEmpty()){
+            throw new ResourceNotFound("user with id " +userid+ " doesn't exist");
+        }
+        else if (booksService.getBooksById(bookid).isEmpty()){
+            throw new ResourceNotFound("no books with the id " +bookid+ " found");
+        }
         return reviewsDAO.addReview(userid, bookid, reviews);
     }
 
     public int editReview (int id, Reviews reviews){
+        if(reviewsDAO.getReviewById(id).isEmpty()){
+            throw new ResourceNotFound("review doesn't exist");
+        }
         return reviewsDAO.editReview(id, reviews);
     }
 
     public int deleteReview (int id){
+        if(reviewsDAO.getReviewById(id).isEmpty()){
+            throw new ResourceNotFound("review doesn't exist");
+        }
         return reviewsDAO.deleteReview(id);
     }
 }
